requirements.txt
# Core dependencies
torch==2.2.0
torchvision==0.17.0
numpy==1.24.3

# Image processing and augmentation
Pillow==10.2.0
opencv-python==4.9.0.80
albumentations==1.3.1

# Data science
pandas==2.1.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Deep learning utilities
timm==0.9.16
tqdm==4.66.1

# Web app
streamlit==1.30.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Optional: Model optimization
# onnx==1.15.0
# onnxruntime==1.16.3
